GerritEnvironment: Make git_multimail callable as Gerrit ref-updated hook
git_multimail has the ability to be directly called as either a git update
or git post-receive hook.  Add the ability for git_multimail to be called
as a Gerrit ref-updated hook (Gerrit does not implement the standard git
hooks), and provide a GerritEnvironment (similar to GitoliteEnvironment
or GenericEnvironment) with relevant defaults.

Gerrit users are probably slightly less likely to use git_multimail since
gerrit already has a built-in email capability.  However:

  * Gerrit's emails appear to only trigger for references submitted as CRs
    (e.g. pushed to the magic refspec refs/for/), not for commits directly
    pushed to the relevant branch.

  * Some sites may have multiple git servers installed (e.g. Gitolite,
    Gerrit, Atlassian Stash) and want similar email styles used regardless
    of which system the repository is hosted under.

The stuff gerrit stores under refs/meta/config is also an implementational
detail of gerrit, which stores access control information.  Users aren't
familiar with the file format, and it is displayed in the Gerrit UI in a
different format.  Users probably don't want to be emailed when access
controls for a project are modified (that's admin-privileged information,
and they have a different UI for viewing it), so just filter this stuff
out.

Signed-off-by: Elijah Newren <newren@palantir.com>