Allow sending emails only for a subset of refs
Many projects use a pull-request model with a central repository, where
new changes are pushed to branches named e.g. feature/whizbang.  However,
users really don't view these branches as "part of the repository" until
they are merged into a 'real' branch, such as master.  And lots of users
are very annoyed at getting email notifications for pushes to these
branches; they only want to be notified when some subset of branches are
updated.

There's a slight complication, though: traditionally git-multimail has
only sent out commit emails for commits that are entirely new to the
repository, but this doesn't play well with this filtering feature.  So,
we need to instead have git-multimail only send out commit emails for
commits that are entirely new to the *relevant subset of the* repository.

Add multimailhook.refFilterInclusionRegex and
multimailhook.refFilterExclusionRegex configuration options, with related
--ref-filter-inclusion-regex and --ref-filter-exclusion-regex command
line options, which can be used to configure which refs are considered
relevant.  Mark ^refs/notes/ as irrelevant by default as git-multimail
emails make no sense for git notes.

We set the default for this in the base Environment to "^refs/notes/".
An empty default would make sense if git_multimail had a specialized
formatting for commits made by git notes, but it currently doesn't.  In
particular, the commit messages of git notes are essentially meaningless
and "filenames" in git notes commits are merely an implementational
detail.  As such, sending emails for these commits, with our current
format, is suboptimal.

Notes are not the only reason for an exclusion, though.  For example,
Gerrit uses refs/changes/ to store commits that have not yet been vetted.
Users think of commits stored under refs/changes/ as being new to the
repository only when they are finally merged into a branch under
refs/heads/.  (There are a variety of reasons for this: (a) these commits
haven't undergone code review and thus are viewed as not yet having been
accepted, (b) Gerrit has a separate mechanism to handle things pushed to
refs/changes/ (emailing just a few relevant folks with a message along
the lines of "You have a Change Review that you need to look at; look at
this nifty web page to add your comments."), (c) only a small subset of
users are interested in any particular refs/changes/ push, whereas
hundreds of developers want to see what enters refs/heads/, (d), gerrit's
standard ref-updated hook, which is similar to git's update or
post-receive hooks, is not invoked for these refs.  It all adds up to
reasons to ignore this particular ref namespace.)  Another example is
Gerrit's refs/cache-automerge/, which is essentially just an
implementational detail that we clearly don't want to spam users about.

Rebase on master, style fixes, code simplifications by:
  Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Elijah Newren <newren@palantir.com>