make Environment available to Push class
Currently, the Push class contains a list of Changes, each of them
pointing to the Environment, hence Push classes can usually access it
saying change.environment, but some methods do not have a change at hand.

We need to access the environment in the computation of new objects to
implement "send emails for only a subset of refs", hence make the
environment directly available.

Signed-off-by: Matthieu Moy <Matthieu.Moy@imag.fr>